Linux compatibility support
close #319

+++ Kainote/Demux.h
@@ -37,7 +37,7 @@ public:
 	bool SaveFont(int i, const wxString& path, wxZipOutputStream* zip = nullptr);
 
 private:
-	static int __stdcall GetSubtitles(long long Start, long long Duration, long long Total, const char* Line, void* ICPrivate);
+	static int __stdcall GetSubtitles(int64_t Start, int64_t Duration, int64_t Total, const char* Line, void* ICPrivate);
 	FFMS_Indexer* indexer = nullptr;
 	std::vector<FFMS_Attachment*> attachments;
 	ProgressSink* progress = nullptr;
